Rename the project folder
