+++ application/app.py
@@ -1,7 +1,7 @@
 import sys
 from PySide2.QtWidgets import QApplication
 
-from splitnotes2.ui.main_window import MainWindow
+from splitguides.ui.main_window import MainWindow
 
 if __name__ == "__main__":
     app = QApplication(sys.argv)
