+++ src/splitguides/ui/hotkey_manager.py
