+++ src/splitguides/ui/layouts/main_window.ui
