+++ src/splitguides/ui/main_window.py
