+++ tests/test_livesplit_client/test_networking.py
@@ -3,7 +3,7 @@ from unittest.mock import patch, MagicMock
 
 import pytest
 
-from splitnotes2.livesplit_client.networking import LivesplitConnection, BUFFER_SIZE
+from splitguides.livesplit_client.networking import LivesplitConnection, BUFFER_SIZE
 
 
 def test_init():
